backend/requirements.txt
django==4.0
djangorestframework==3.14.0
decouple~=0.0.7
environ~=1.0
django-cors-headers==4.3.1
drf-yasg==1.21.7
django-rest-passwordreset==1.3.0
django-storages==1.14.2
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.3.1
pillow==10.2.0
boto3==1.18.36
django-environ==0.11.2
asgiref==3.7.2
botocore==1.21.65
gunicorn==21.2.0
inflection==0.5.1
jmespath==0.10.0
packaging==23.2
python-decouple==3.8
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==1.26.18